feat: Add global configuration module
This commit adds a new module called `config.rs` which introduces a global configuration for the application. The `Config` struct provides methods for retrieving and setting the node address and mining address. It also includes a lazy static instance of the `Config` struct called `GLOBAL_CONFIG`. This module aims to centralize the configuration settings and make them easily accessible throughout the application.